test: Remove intermittent and presumed fixed tsan race suppressions
LoadWallet was added in commit 4ddbddea1889d7a32470f21ad97e132e71c3e8c1, which
points to a traceback with BerkeleyBatch in it. Now that BDB is removed,
this can be removed as well.

The race in DatabaseBatch was added in commit
244b063b9fdb61c853b6fb55c2d5aa3fa0bbdd85, which does not point to a
traceback. This was likely also fixed with the BDB removal.

If not, the suppressions should be added back, mentioning that they are
intermittent and including a traceback and possibly steps to reproduce.